+++ src/components/input/Input.stories.ts
@@ -1,0 +1,44 @@
+import { boolean, text, select } from '@storybook/addon-knobs';
+import { action } from '@storybook/addon-actions';
+import Vue, { PropType } from 'vue';
+import WvuiInput from './Input.vue';
+import { InputType } from './InputType';
+
+export default {
+	title: 'Components/Input',
+	component: WvuiInput,
+	parameters: { layout: 'centered' }
+};
+
+export const configurable = (): Vue.Component =>
+	Vue.extend( {
+		components: { WvuiInput },
+		props: {
+			disabled: { type: Boolean, default: boolean( 'Disabled', false ) },
+			type: {
+				type: String as PropType<keyof typeof InputType>,
+				default: select( 'Input Type', Object.keys( InputType ), 'Search' )
+			},
+			placeholder: { type: String, default: text( 'Placeholder', 'Search...' ) }
+		},
+		data() {
+			return { InputType };
+		},
+		methods: {
+			input: action( 'input' ),
+			change: action( 'change' ),
+			focus: action( 'focus' ),
+			blur: action( 'blur' )
+		},
+		template: `
+		<wvui-input
+			:placeholder="placeholder"
+			:type="InputType[ type ]"
+			:disabled="disabled"
+			@input="input"
+			@change="change"
+			@focus="focus"
+			@blur="blur"
+		></wvui-input>
+		`
+	} );