+++ src/components/input/Input.test.ts
@@ -1,0 +1,46 @@
+import { mount, shallowMount } from '@vue/test-utils';
+import WvuiInput from './Input.vue';
+import { InputType } from './InputType';
+
+describe( 'matches the snapshot', () => {
+	type Case = [string, Record<keyof unknown, unknown>];
+
+	const cases: Case[] = Object.values( InputType ).map( ( type ) => [
+		`Input Type: ${type} `,
+		{ type }
+	] ) as Case[];
+
+	test.each( cases )( 'Case %# %s: (%p) => HTML', ( _, props ) => {
+		const wrapper = mount( WvuiInput, { propsData: props } );
+
+		expect( wrapper.element ).toMatchSnapshot();
+	} );
+} );
+
+it( 'emits input events', () => {
+	const wrapper = shallowMount( WvuiInput );
+
+	wrapper.get( 'input' ).trigger( 'input' );
+	expect( wrapper.emitted().input ).toBeTruthy();
+} );
+
+it( 'emits change events', () => {
+	const wrapper = shallowMount( WvuiInput );
+
+	wrapper.get( 'input' ).trigger( 'change' );
+	expect( wrapper.emitted().change ).toBeTruthy();
+} );
+
+it( 'emits focus events', () => {
+	const wrapper = shallowMount( WvuiInput );
+
+	wrapper.get( 'input' ).trigger( 'focus' );
+	expect( wrapper.emitted().focus ).toBeTruthy();
+} );
+
+it( 'emits blur events', () => {
+	const wrapper = shallowMount( WvuiInput );
+
+	wrapper.get( 'input' ).trigger( 'blur' );
+	expect( wrapper.emitted().blur ).toBeTruthy();
+} );