+++ src/components/input/InputType.test.ts
@@ -1,0 +1,19 @@
+import { InputType, isInputType } from './InputType';
+
+describe( 'isInputType()', () => {
+	// [description, input, expected]
+	type Case = [string, string, boolean];
+
+	const cases: Case[] = [
+		...Object.values( InputType ).map(
+			( type ) => [ type, type, true ]
+		) as Case[],
+		[ 'invalid', 'invalid', false ]
+	];
+
+	test.each( cases )(
+		'Case %# %s: (%p) => %p',
+		( _, action, expected ) =>
+			expect( isInputType( action ) ).toStrictEqual( expected )
+	);
+} );