+++ src/entries/wvui.ts
@@ -1,5 +1,8 @@
 import WvuiButton from '@/components/button/Button.vue';
+import WvuiInput from '@/components/input/Input.vue';
 
 export default {
-	WvuiButton
+	version: VERSION,
+	WvuiButton,
+	WvuiInput
 };