+++ src/types/VueLoader.d.ts
@@ -1,5 +1,5 @@
 // Assume every file with a .vue extension is of type Vue.
 declare module '*.vue' {
-    import Vue from 'vue';
-    export default Vue;
+	import Vue from 'vue';
+	export default Vue;
 }