+++ webpack.config.js
@@ -9,6 +9,7 @@ const { VueLoaderPlugin } = require( 'vue-loader' );
 const webpack = require( 'webpack' );
 const path = require( 'path' );
 const { BundleAnalyzerPlugin } = require( 'webpack-bundle-analyzer' );
+const { version } = require( './package.json' );
 
 const resolve = {
 	extensions: [ '.js', '.ts' ],
@@ -16,8 +17,13 @@ const resolve = {
 	alias: { '@': path.resolve( __dirname, './src' ) }
 };
 
-const jsSourceMapExtention = '.map.json';
+// The extension used for source map files. Per T173491, files with a .map extension cannot be
+// served from prod. It doesn't seem to be practical to rename the CSS source maps.
+const jsSourceMapExtension = '.map.json';
 
+// Enumeration of chunk names. The key is a symbol and the value is the chunk name and file stem.
+// # See readme.md#different-builds for details. Some of these chunks have specific entry points
+// under src/entries and others are generated automatically.
 const Chunk = {
 	Wvui: 'wvui'
 };
@@ -37,10 +43,14 @@ function rules( mode ) {
 			// Do not process node_modules at all. This means no transpilation of dependencies.
 			include: path.resolve( __dirname, 'src' ),
 			use: {
-				// Type checking is performed by ForkTsCheckerWebpackPlugin.
 				loader: 'ts-loader',
 				options: {
-					transpileOnly: true
+					// Type checking is performed by ForkTsCheckerWebpackPlugin which does not emit
+					// TypeScript definitions. See
+					// https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/issues/49
+					transpileOnly: mode === 'development',
+					// Needed when transpiling.
+					appendTsSuffixTo: [ /\.vue$/ ]
 				}
 			}
 		},
@@ -71,12 +81,17 @@ function rules( mode ) {
  * List of webpack plugins to be used
  * in common configuration
  *
+ * @param {'development' | 'production' | 'none'} [mode]
  * @return {webpack.Plugin[]}
  * */
-function plugins() {
+function plugins( mode ) {
 	return [
-		new ForkTsCheckerWebpackPlugin(),
-		new MiniCssExtractPlugin()
+		...( mode === 'development' ? [ new ForkTsCheckerWebpackPlugin() ] : [] ),
+		new MiniCssExtractPlugin(),
+		// The DefinePlugins entries should be kept in sync with Environment.d.ts.
+		new webpack.DefinePlugin( {
+			VERSION: JSON.stringify( version )
+		} )
 	];
 }
 
@@ -107,7 +122,7 @@ module.exports = ( _env, argv ) => ( {
 
 	performance: {
 		// The default filter excludes map files but we rename ours. See T173491.
-		assetFilter: ( filename ) => !filename.endsWith( jsSourceMapExtention )
+		assetFilter: ( filename ) => !filename.endsWith( jsSourceMapExtension )
 	},
 
 	// Accurate source maps come at the expense of build time. The source map is intentionally
@@ -140,7 +155,7 @@ module.exports = ( _env, argv ) => ( {
 	},
 
 	output: {
-		sourceMapFilename: `[file]${jsSourceMapExtention}`,
+		sourceMapFilename: `[file]${jsSourceMapExtension}`,
 		// Set the name to avoid possible Webpack runtime collisions of globals with other Webpack
 		// runtimes. See https://webpack.js.org/configuration/output/#outputuniquename.
 		library: 'wvui',
@@ -163,7 +178,7 @@ module.exports = ( _env, argv ) => ( {
 			// Don't delete the ES5 linter config.
 			cleanOnceBeforeBuildPatterns: [ '**/*', '!.eslintrc.json' ]
 		} ),
-		...plugins(),
+		...plugins( argv.mode ),
 		new VueLoaderPlugin(),
 		new BundleAnalyzerPlugin( {
 			analyzerMode: argv.mode === 'development' ? 'disabled' : 'static',