+++ changelog.md
@@ -6,7 +6,13 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 -   [update][icon] Add new component and styles
 -   [update][button] Add initial styles
+-   [build][fix] Publish bundled CSS and types
+-   [build][dev] Include WVUI version in release
+-   [build][dev] Move docs to development release script
 -   [update][button] Flesh out props
+-   [dev] Upgrade dependencies
+-   [build][fix] Export TypeScript definitions
+-   [update] Add initial TextInput styles and props
 -   [build][dev] Enable development releases
 -   [dev] Add wikimedia-ui theme
 -   [dev] Prevent prettier from checking less files